Use dependency injection everywhere, move services to the correct source set
