+++ app/src/google/kotlin/com/keylesspalace/tusky/di/UpdateCheckModule.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2023 Tusky Contributors
+ *
+ * This file is a part of Tusky.
+ *
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation; either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * Tusky is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with Tusky; if not,
+ * see <http://www.gnu.org/licenses>.
+ */
+
+package com.keylesspalace.tusky.di
+
+import android.content.Context
+import com.google.android.play.core.appupdate.AppUpdateManager
+import com.google.android.play.core.appupdate.AppUpdateManagerFactory
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module
+class UpdateCheckModule {
+    @Provides
+    @Singleton
+    fun providesAppUpdateManager(
+        context: Context
+    ): AppUpdateManager = AppUpdateManagerFactory.create(context)
+}
